feat(executor-agent-rs): add V2 protocol with handle-based process multiplexing
V2 protocol (selected via EXECUTOR_PROTOCOL=v2 or --protocol v2, default
remains v1) introduces:

- spawn: unified exec+shell — PTY and stdin are independent flags
- Handle-based process addressing (proc-{uuid8}), no PID exposure
- Multiplexed connections: multiple processes per connection, no hijacking
- Server-push events: stdout/stderr/exit sent proactively with handle tag
- stdin support for non-PTY processes
- stat and list_dir: file metadata without spawning processes
- Structured errors with code + message
- Timeout with SIGKILL escalation

V1 protocol is unchanged and fully operational (verified 11/11 on kind).

write_file chunked upload returns NOT_IMPLEMENTED pending a state-machine
reader that can interleave file chunks with other requests.

+++ rust/executor-agent/src/pty_util.rs
@@ -44,8 +44,7 @@ pub fn setup_slave_terminal(slave_fd: i32) -> io::Result<()> {
         .map_err(|e| io::Error::other(format!("setsid: {}", e)))?;
 
     // Set controlling terminal
-    #[allow(clippy::unnecessary_cast)]
-    let ret = unsafe { libc::ioctl(slave_fd, libc::TIOCSCTTY as libc::c_int, 0i32) };
+    let ret = unsafe { libc::ioctl(slave_fd, libc::TIOCSCTTY, 0i32) };
     if ret < 0 {
         return Err(io::Error::last_os_error());
     }
